feat(amuxd): P4+P5 停发偏好字段并删除 daemon MRU
ADR-0007 的收尾两步一起做。amuxd 从此只持有模型**能力**（catalog），不再持有
任何模型**偏好**。

P4 停止填充两个 proto 字段（定义保留，老客户端仍能解析，只是恒为空）：

- `ActorPresence.default_model`(12) —— daemon MRU 表头
- `ActorPresence.worktrees`(7)

两个必须一起停。`worktrees[0].default_model`（WorktreeCatalog 字段 3）是同一个
偏好的第二条通路，两端解析链都写成 `presence.defaultModel || legacy?.defaultModel`
（runtime-state-store.ts:312-313、SessionListViewModel.swift:429-432），只停 12
号偏好照样从 7 号流进来。

`RuntimeInfo.current_model` 不在此列 —— 计划原文列了它，但 ADR-0004 phase 7
已于 8/4 停发 `runtime/{id}/state`，整条消息今天没有发布者，无从停发。retain 上
真正承载模型的是 `LiveSession.current_model`（manager.rs:1489），**保留**：它是
P3 客户端解析链的 live 那一层，也是 ADR-0007「daemon 仍在进程内跟踪当前模型」
那条例外的出口。

`catalog_models` 一行未动。这是关键区分：目录是能力，偏好才是搬走的东西。远端
客户端只有 retain 一条路拿目录，iOS 更是连 loopback 兜底都没有；把它一起停掉
等于让远端没有模型可选。

P5 删除：

- `config/model_mru.rs` 整个模块，以及 `~/.amuxd/cache/model-mru.toml`（不迁移）
- manager 的 `recent_models` / `learn_session_model` / `set_current_model` 里的
  MRU 写入。`set_current_model` 本身保留 —— 它仍写 `session_participants.model`
- `BackendCatalog.recent_models` 与 `attach_recent_models`
- `supervisor.recent_catalog_models`
- 客户端的 `recentModels` 通道与 `firstAvailableRecentModel`

`first_available` 没跟着删 —— 它不是 MRU 专属，是「候选里挑目录仍提供的那个」
这条通用校验，三个后端都在用。移到 `config/model_resolution.rs`。

`learn_session_model` 一删，它独占的 `AgentBackend::session_model` 就没有调用方
了，连同三个后端的实现一并删除（opencode_http / claude_agent / cursor_sdk，pi
用 trait 默认实现）。每一处的注释都写着自己存在的理由是 "teach the device MRU"。
`client::session_model_id` 保留：同名但服务的是 resume 路径
（opencode_http/mod.rs:755，重新挂上已有会话时读回它上次实际跑在什么模型上），
那是关于这一个会话的事实，不是设备级偏好。

cron 存量回填一并退役：它读的正是这里删掉的 daemon MRU，其存在意义就是 P5 之前
那个窗口。留着只会永远不结论、每次启动重试。

**放行门槛已豁免**。计划原定等 `amux.actor_client_versions` 显示无低于 P3 版本
的活跃 reporter 再执行；本次按决策直接发。执行时 P3（#920）尚未发版，线上 P3
客户端为零，近 7 日活跃 tauri 13 / daemon 13 / expo 4 / ios 2 台全部低于门槛，
其模型 pill 会退回各自兜底。这是明知代价的选择。计划文档 P4 一节已如实记录，
并补上原表漏掉的第四个 client_type（expo）。

验证：cargo test -p amuxd 9 个 suite / 5276 passed / 0 failed / 1 ignored（缺
opencode 二进制的既有用例）；pnpm typecheck 0；pnpm lint 0 errors；app 单测
438 文件 / 2820 passed。
